その他のコマンドをmiscパッケージに
