+++ packages/misc/commands/discordinfo.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder, ChannelType } = require('discord.js');
-const { LANG, strFormat } = require('../util/languages');
+const { LANG, strFormat } = require('../../../util/languages');
 
 module.exports = {
 	data: new SlashCommandBuilder()
