+++ packages/misc/commands/randomnum.js
@@ -1,7 +1,7 @@
 // @ts-check
 
-const { LANG, strFormat } = require('../util/languages');
-const { SimpleSlashCommandBuilder } = require('../common/SimpleCommand');
+const { LANG, strFormat } = require('../../../util/languages');
+const { SimpleSlashCommandBuilder } = require('../../../common/SimpleCommand');
 
 const DEFAULT_MIN_VALUE = 0;
 const DEFAULT_MAX_VALUE = 99;
