+++ packages/misc/commands/receiveupdate.js
@@ -5,7 +5,7 @@ const {
 	PermissionsBitField,
 	NewsChannel,
 } = require('discord.js');
-const { LANG, strFormat } = require('../util/languages');
+const { LANG, strFormat } = require('../../../util/languages');
 
 module.exports = {
 	data: new SlashCommandBuilder()
