+++ packages/misc/commands/testx.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require('discord.js');
-const { LANG } = require('../util/languages');
+const { LANG } = require('../../../util/languages');
 
 module.exports = {
 	data: new SlashCommandBuilder()
