+++ packages/misc/package.json
@@ -1,0 +1,12 @@
+{
+	"name": "misc",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"devDependencies": {},
+	"scripts": {
+		"test": "echo \"Error: no test specified\" && exit 1"
+	},
+	"author": "",
+	"license": "GPL-3.0-only"
+}
